added negotiated base class loading (highest base version loads first). still needed: * refactor acitivation by adding a HelloInstall class and removing install from Hello class which is dependent on base class. * add framework versions to wp_options table to track plugin framework versions, to be used on loading plugins to find highest available versions.
Decision was made to abandon this branch, and instead pursue a simpler method where an updated Simpli base class library would just be renamed to Simpli2,Simpli3, etc. and be deployed along with the rest of the framework which would follow a different versioning scheme.  This promises to be a much more bug-free/robust approach if not as cool.