Refactored code to simplify plugin.php, and class library organization.
Moved class libraries to within subdirectories whose parent directories are namespaces of 'Simpli' (in the case of the framework) and the slug prefix (in the case of the Plugin code, which is 'Simpli' for the demo plugin 'Simpli_Hello').

Added some properties to the Plugin base classes ( framework_version,text_domain)
Refactored Logger class to behave like other classes and moved instantiation to the extended Plugin class.
Moved much of the plugin.php code to the extended Plugin::init method.

Next Step: Add a static framework class to simplify plugin.php code and to handle the plugin loading.